Move servant types into separate package
