+++ smash/smash.cabal
@@ -48,10 +48,8 @@ library
       Cardano.SMASH.DBSync.Db.Database
 
   exposed-modules:
-    Cardano.SMASH.API
     Cardano.SMASH.DB
     Cardano.SMASH.DBSync.Db.Delete
-    Cardano.SMASH.DBSync.Db.Error
     Cardano.SMASH.DBSync.Db.Insert
     Cardano.SMASH.DBSync.Db.Migration
     Cardano.SMASH.DBSync.Db.Migration.Haskell
@@ -60,13 +58,11 @@ library
     Cardano.SMASH.DBSync.Db.Query
     Cardano.SMASH.DBSync.Db.Run
     Cardano.SMASH.DBSync.Db.Schema
-    Cardano.SMASH.DBSync.Db.Types
     Cardano.SMASH.DBSync.Metrics
     Cardano.SMASH.FetchQueue
     Cardano.SMASH.FetchQueue.Retry
     Cardano.SMASH.Lib
     Cardano.SMASH.Offline
-    Cardano.SMASH.Types
 
   other-modules:      Paths_smash
   hs-source-dirs:     src
@@ -118,6 +114,7 @@ library
     , servant-server
     , servant-swagger
     , shelley-spec-ledger
+    , smash-servant-types
     , swagger2
     , template-haskell
     , text
@@ -149,6 +146,7 @@ executable smash-exe
     , cardano-slotting
     , optparse-applicative
     , smash
+    , smash-servant-types
 
   default-language:   Haskell2010
   default-extensions:
